+++ src/rust/ensogl/lib/core/src/animation.rs
@@ -1,11 +1,14 @@
 //! This module provides different kind of animation utilities, such as physics based animation and
 //! animation with easing functions.
 
-pub mod physics;
-mod loops;
 pub mod easing;
+pub mod physics;
+
+mod frp;
+mod loops;
 
 pub use loops::*;
+pub use frp::*;
