+++ src/rust/ide/view/graph-editor/src/component/node/output/area.rs
@@ -5,27 +5,17 @@ use ensogl::display::traits::*;
 
 use enso_frp as frp;
 use enso_frp;
+use ensogl::Animation;
+use ensogl::Tween;
+use ensogl::application::Application;
 use ensogl::data::color;
 use ensogl::display::scene::Scene;
-use ensogl::display::shape::AnyShape;
-use ensogl::display::shape::BottomHalfPlane;
-use ensogl::display::shape::Circle;
-use ensogl::display::shape::PixelDistance;
-use ensogl::display::shape::Pixels;
 use ensogl::display::shape::StyleWatch;
-use ensogl::display::shape::Rect;
-use ensogl::display::shape::Var;
-use ensogl::display::shape::primitive::def::class::ShapeOps;
 use ensogl::display;
-use ensogl::gui::component::DEPRECATED_Animation;
 use ensogl::gui::component;
+use ensogl_text as text;
 use ensogl_theme as theme;
 use span_tree;
-use ensogl::application::Application;
-use ensogl_text as text;
-use ensogl_text::buffer::data::unit::traits::*;
-use ensogl::gui::component::Animation;
-use ensogl::gui::component::Tween;
 
 use crate::Type;
 use crate::component::node;
