+++ src/rust/ide/view/graph-editor/src/component/visualization/container.rs
@@ -25,7 +25,7 @@ use ensogl::display::scene::Scene;
 use ensogl::display::shape::*;
 use ensogl::display::traits::*;
 use ensogl::display;
-use ensogl::gui::component::DEPRECATED_Animation;
+use ensogl::DEPRECATED_Animation;
 use ensogl::application::Application;
 use ensogl::gui::component;
 use ensogl::system::web;
