+++ src/rust/ide/view/src/searcher.rs
@@ -11,7 +11,7 @@ use enso_frp as frp;
 use ensogl::application;
 use ensogl::application::{Application, shortcut};
 use ensogl::display;
-use ensogl::gui::component::DEPRECATED_Animation;
+use ensogl::DEPRECATED_Animation;
 use ensogl_gui_components::list_view;
 use ensogl_gui_components::list_view::ListView;
